chore(model): Move files around and split big files into smaller ones
